Moved test apis out of server.js and into test routes.
